+++ include/type.h
@@ -6,7 +6,7 @@
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/29 11:17:36 by tkuramot          #+#    #+#             */
-/*   Updated: 2023/10/04 03:25:44 by tsishika         ###   ########.fr       */
+/*   Updated: 2023/10/11 11:11:58 by tsishika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
